--- src/main/kotlin/reverseString/ReverseString.kt
@@ -1,20 +1,0 @@
-package reverseString
-
-class ReverseString {
-    private var tmpChar: Char = '\u0000'
-
-    fun reverseString(s: CharArray): CharArray {
-        val length = s.size
-        val halfLength = length / 2
-        for (i in 0 until halfLength) {
-            swap(s, i, length - 1 - i)
-        }
-        return s
-    }
-
-    private fun swap(s: CharArray, i: Int, j: Int) {
-        tmpChar = s[i]
-        s[i] = s[j]
-        s[j] = tmpChar
-    }
-}
